// prisma/generated/zod/inputTypeSchemas/UserOrganizationRelationCountOrderByAggregateInputSchema.ts
import type { Prisma } from '@prisma/client';

import { z } from 'zod';
import { SortOrderSchema } from './SortOrderSchema';

export const UserOrganizationRelationCountOrderByAggregateInputSchema: z.ZodType<Prisma.UserOrganizationRelationCountOrderByAggregateInput> = z.object({
  userId: z.lazy(() => SortOrderSchema).optional(),
  organizationId: z.lazy(() => SortOrderSchema).optional(),
  role: z.lazy(() => SortOrderSchema).optional(),
  createdAt: z.lazy(() => SortOrderSchema).optional()
}).strict();

export default UserOrganizationRelationCountOrderByAggregateInputSchema;